[ci] Run Mac MSBuild tests on 6 nodes instead of 3.
