+++ src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/AndroidDependenciesTests.cs
@@ -9,7 +9,7 @@ using Xamarin.ProjectTools;
 namespace Xamarin.Android.Build.Tests
 {
 	[TestFixture]
-	[Category ("Node-3")]
+	[Category ("Node-2")]
 	[Parallelizable (ParallelScope.Children)]
 	public class AndroidDependenciesTests : BaseTest
 	{
