+++ src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/PackagingTest.cs
@@ -12,7 +12,7 @@ using Xamarin.Android.Tools;
 
 namespace Xamarin.Android.Build.Tests
 {
-	[Category ("Node-3")]
+	[Category ("Node-6")]
 	[Parallelizable (ParallelScope.Children)]
 	public class PackagingTest : BaseTest
 	{
