+++ src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/Tasks/ManagedResourceParserTests.cs
@@ -13,7 +13,7 @@ using System.Text.RegularExpressions;
 
 namespace Xamarin.Android.Build.Tests {
 	[TestFixture]
-	[Category ("Node-2")]
+	[Category ("Node-5")]
 	[Parallelizable (ParallelScope.Children)]
 	public class ManagedResourceParserTests : BaseTest {
 		const string ValuesXml = @"<?xml version=""1.0"" encoding=""utf-8""?>
