+++ src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/Tasks/NdkUtilTests.cs
@@ -13,7 +13,7 @@ using Xamarin.ProjectTools;
 namespace Xamarin.Android.Build.Tests.Tasks {
 
 	[TestFixture]
-	[Category ("Node-2")]
+	[Category ("Node-5")]
 	public class NdkUtilTests : BaseTest {
 
 		List<BuildErrorEventArgs> errors;
