+++ src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/Tasks/ResolveMonoAndroidSdksTests.cs
@@ -7,7 +7,7 @@ using Xamarin.ProjectTools;
 namespace Xamarin.Android.Build.Tests
 {
 	[TestFixture]
-	[Category ("Node-2")]
+	[Category ("Node-5")]
 	public class ResolveMonoAndroidSdksTests : BaseTest
 	{
 		static readonly string [] parameters = new [] { "_ResolveMonoAndroidSdksDependsOn=" };
