+++ src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/Tasks/ValidateJavaVersionTests.cs
@@ -10,7 +10,7 @@ using ValidateJavaVersion = Xamarin.Android.Tasks.Legacy.ValidateJavaVersion;
 namespace Xamarin.Android.Build.Tests
 {
 	[TestFixture]
-	[Category ("Node-2")]
+	[Category ("Node-5")]
 	public class ValidateJavaVersionTests : BaseTest
 	{
 		string path;
