+++ src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/XASdkTests.cs
@@ -15,7 +15,7 @@ namespace Xamarin.Android.Build.Tests
 {
 	[TestFixture]
 	[NonParallelizable] // On MacOS, parallel /restore causes issues
-	[Category ("Node-2")]
+	[Category ("Node-5")]
 	public class XASdkTests : BaseTest
 	{
 		/// <summary>
